refactor: Move sidebar_data into separate file
